Rename some subsections in planner
